# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pharma Order Management API",
    "version": "1.0.0",
    "description": "API para gestionar tiendas, sucursales, usuarios y órdenes en el sistema de farmacia."
  },
  "servers": [
    {
      "url": "http://localhost:3010",
      "description": "Local dev server"
    }
  ],
  "paths": {
...
    "/api/categories": {
      "get": {
        "summary": "Listar categorías de productos",
        "tags": [
          "Product Categories"
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías de productos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear categoría de productos",
        "tags": [
          "Product Categories"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoría creada"
          }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "summary": "Listar productos",
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "Lista de productos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear producto",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "decimal"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "ProductCreateInput": {
        "type": "object",
        "required": [
          "name",
          "price",
          "categoryId"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "example": 150.75
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "CategoryCreateInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "example": "Categoría ABC"
          }
        }
      },
...
    }
  }
}